refactor p5 related methods to lib file
