+++ src/components/Editor.jsx
@@ -1,7 +1,7 @@
 import CodeMirror from '@uiw/react-codemirror'
 import { clojure } from '@nextjournal/lang-clojure'
 import { useEffect, useState } from 'react'
-import { clearP5import, clearWindowGlobals, compileAndSet, importP5, removeDefaultCanvas } from '../utils'
+import { clearP5import, clearWindowGlobals, compileAndSet, importP5, removeDefaultCanvas } from '../lib/p5'
 import { useSearchParams } from 'react-router-dom'
 import { LZString } from '../LZString'
